Use unionId instead of openId
Unlike the openId the unionId is stable accross applications.

As often one application has several WeChat applications e.g. for Android, iOS and Web
the unionId should get used otherwise a user which signed up on e.g. Android can not signup
with the same WeChat account on Web or iOS as it would have a different openID for each
of these applications.

> OpenID is an unique encrypted WeChat ID for each user of an OA (including Mini Program).

> UnionID is an unique ID to identify a single user under the same WeChat open platform account.
If a developer possesses multiple mobile applications, website applications, and OAs (including Mini Programs),
they can use a UnionID to differentiate the uniqueness of users.
In other words, UnionID offers a mechanism for developers to acquire a user’s data from multiple
platforms and attribute actions back to the specific user.

https://wiredcraft.com/blog/wechat-for-dummies-user-data-and-authorization/
https://stackoverflow.com/questions/40044572/wechat-open-platform-openid-and-unionid

This change would break existing users of the strategy as we change the `uid` attribute.
However, I don't think there is a way to make this backward compatible :thinking:.