Add image field to user info hash
This information is meant to be as normalized as possible, so the schema below will be filled to the greatest degree available given the provider upon authentication. Fields marked required will always be present. Note that this guide may also be useful to developers implementing custom strategies to know how to provide results.

https://github.com/omniauth/omniauth/wiki/Auth-Hash-Schema